make icon its own module
abstract the method to make icon markup into its own module.

this allows for it to be used in JavaScript environment that does not
have access to nodeism like `fs` or `__dirname` (such as browserify or
webpack).

this method could potentially be re-used in the browser asset
bundle, as it is the same.